Agrego correr con interprete ruby

+++ app/db/migrate/20250122005411_submission.rb
@@ -10,7 +10,7 @@ class Submission < ActiveRecord::Migration[7.2]
       t.text :interpreter_output
 
       t.integer :memory_used
-      t.integer :time_used
+      t.float :time_used
 
       t.string :status, null: false, default: "queued"
